Revert "added skipping Git tests on windows machines, because the required pear package is incompatible with windows"
This reverts commit 9c2f9c851f7d9d057151d9d73033b93d585767e1.

+++ test/classes/phing/tasks/ext/GitTasks/GitTestsHelper.php
@@ -1,24 +1,14 @@
 <?php
-
 class GitTestsHelper
 {
-    public static function rmdir($dir)
+    public static function rmdir($dir) 
     {
-        if (! file_exists($dir)) {
-            return true;
-        }
-        if (! is_dir($dir)) {
-            return unlink($dir);
-        }
+        if (!file_exists($dir)) return true;
+        if (!is_dir($dir)) return unlink($dir);
         foreach (scandir($dir) as $item) {
-            if ($item == '.' || $item == '..') {
-                continue;
-            }
-            if (! self::rmdir($dir . DIRECTORY_SEPARATOR . $item)) {
-                return false;
-            }
+            if ($item == '.' || $item == '..') continue;
+            if (!self::rmdir($dir.DIRECTORY_SEPARATOR.$item)) return false;
         }
-
         return rmdir($dir);
     }
 
@@ -26,33 +16,35 @@ class GitTestsHelper
      * Get relative date
      *
      * @param int $timestamp Timestamp to us as pin-point
+     * @param string $type Whether 'fulldate' or 'time'
      */
-    public static function getRelativeDate($timestamp)
+    public static function getRelativeDate($timestamp, $type = 'fulldate')
     {
         // calculate the diffrence
-        $timediff = time() - $timestamp;
+        $timediff = time () - $timestamp ;
 
         if ($timediff < 3600) {
             if ($timediff < 120) {
                 $returndate = "1 minute ago";
             } else {
-                $returndate = ceil($timediff / 60) . " minutes ago";
+                $returndate =  ceil($timediff / 60) . " minutes ago";
             }
-        } elseif ($timediff < 7200) {
+        } else if ($timediff < 7200) {
             $returndate = "1 hour ago.";
-        } elseif ($timediff < 86400) {
+        } else if ($timediff < 86400) {
             $returndate = ceil($timediff / 3600) . " hours ago";
-        } elseif ($timediff < 172800) {
+        } else if ($timediff < 172800) {
             $returndate = "1 day ago.";
-        } elseif ($timediff < 604800) {
+        } else if ($timediff < 604800) {
             $returndate = ceil($timediff / 86400) . " days ago";
-        } elseif ($timediff < 1209600) {
+        } else if ($timediff < 1209600) {
+            $returndate = "1 week ago.";
             $returndate = ceil($timediff / 86400) . " days ago";
-        } elseif ($timediff < 2629744) {
+        } else if ($timediff < 2629744) {
             $returndate = ceil($timediff / 86400) . " days ago";
-        } elseif ($timediff < 3024000) {
+        } else if ($timediff < 3024000) {
             $returndate = ceil($timediff / 604900) . " weeks ago";
-        } elseif ($timediff > 5259486) {
+        } else if ($timediff > 5259486) {
             $returndate = ceil($timediff / 2629744) . " months ago";
         } else {
             $returndate = ceil($timediff / 604900) . " weeks ago";
